requirements.txt
streamlit
pandas
sentence-transformers
faiss-cpu
deep-translator
edge-tts
numpy
gTTS==2.3.0
streamlit==1.30.0
pandas==2.1.0
numpy==1.27.0
gTTS==2.3.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
streamlit-option-menu==0.5.5
torch
torchvision
transformers
torch==2.2.0
torchvision==0.17.0
transformers==5.3.0
